requirements.txt
Django==6.0
Pillow==12.0.0
django-ckeditor==6.7.3
python-decouple==3.8
